pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-chess"
version = "0.1.0"
description = "LLM vs LLM chess match runner using OpenRouter"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "python-chess>=1.999",
  "openai>=1.40.0",
  "stockfish>=3.28.0",
  "flask>=3.0.0",
  "cairosvg>=2.7.0",
  "pillow>=10.3.0",
  "torch>=2.0.0",
]

[project.scripts]
llm-chess = "llm_chess.cli:main"
llm-chess-web = "llm_chess.web:main"
llm-chess-train = "llm_chess.chess_transformer.train:main"
llm-chess-generate = "llm_chess.chess_transformer.generate_games:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
llm_chess = ["web/**/*", "assets/**/*"]

[tool.setuptools.packages.find]
where = ["src"]
